E2e tests added for keyless scenarios

+++ test/e2e_test.go
@@ -117,76 +117,10 @@ func shutdown() {
 	deleteS3Bucket(bucket)
 }
 
-//func TestCodeSignAndVerify(t *testing.T) {
-//	os.Setenv(integrity.ExperimentalEnv, "0")
-//	viper.Set("privatekey", privateKey)
-//	funcDefer, err := mockStdin(t, pass)
-//	if err != nil {
-//		t.Fatal(err)
-//	}
-//	defer funcDefer()
-//
-//	sbo := o.SignBlobOptions{
-//		SignBlobOptions: options.SignBlobOptions{
-//			Base64Output: true,
-//			Registry:     options.RegistryOptions{},
-//		},
-//	}
-//	err = sign.SignAndUploadCode(awsClient, "utils/testing_lambda", &sbo, ro)
-//	if err != nil {
-//		t.Fatal(err)
-//	}
-//
-//	functionArn := initCodeLambda(t)
-//
-//	successTagValue := "Function signed and verified"
-//	success, timeout := findTag(t, functionArn, lambdaClient, "Function clarity result", successTagValue)
-//	if timeout {
-//		t.Fatal("test failed on timout, the required tag not added in the time period")
-//	}
-//	if !success {
-//		t.Fatal("test failure: no " + successTagValue + " tag in the signed function")
-//	}
-//	fmt.Println(successTagValue + " tag found in the signed function")
-//	deleteLambda(codeFuncName)
-//}
-
-func TestImageSignAndVerify(t *testing.T) {
-	viper.Set("privatekey", privateKey)
-	os.Setenv(integrity.ExperimentalEnv, "0")
-	funcDefer, err := mockStdin(t, pass)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer funcDefer()
-
-	ko := options.KeyOpts{KeyRef: privateKey, PassFunc: passFunc}
-	err = s.SignCmd(ro, ko, options.RegistryOptions{}, nil, []string{imageUri}, "", "", true, "", "", "", false, false, "", false)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	functionArn, err := createImageLambda(t)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	successTagValue := "Function signed and verified"
-	success, timeout := findTag(t, functionArn, lambdaClient, "Function clarity result", successTagValue)
-	if timeout {
-		t.Fatal("test failed on timout, the required tag not added in the time period")
-	}
-	if !success {
-		t.Fatal("test failure: no " + successTagValue + " tag in the signed function")
-	}
-	fmt.Println(successTagValue + " tag found in the signed function")
-	deleteLambda(imageFuncName)
-}
-
 func TestCodeSignAndVerifyKeyless(t *testing.T) {
-	viper.Set("privatekey", "")
 	os.Setenv(integrity.ExperimentalEnv, "1")
-	switchConfigurationToKeyless()
+	switchConfiguration(true, "")
+
 	jwt := getEnvVar("jwt_token", "token ID")
 	sbo := o.SignBlobOptions{
 		SignBlobOptions: options.SignBlobOptions{
@@ -215,15 +149,19 @@ func TestCodeSignAndVerifyKeyless(t *testing.T) {
 	}
 	fmt.Println(successTagValue + " tag found in the signed function")
 	deleteLambda(codeFuncName)
+	deleteS3BucketContent(&bucket, []string{"function-clarity.zip"})
 }
 
 func TestCodeImageAndVerifyKeyless(t *testing.T) {
 	viper.Set("privatekey", "")
 	os.Setenv(integrity.ExperimentalEnv, "1")
-	switchConfigurationToKeyless()
-	fmt.Println("testing123")
-	fmt.Println(getEnvVar("jwt_token", "token ID"))
-	jwt := getEnvVar("jwt_token", "token ID")
+	log.Printf("privateKey: %v\n", privateKey)
+	log.Printf("publicKey: %v\n", viper.GetString("publickey"))
+	log.Printf("IsExperimentalEnv: %v\n", integrity.IsExperimentalEnv())
+
+	switchConfiguration(true, "")
+
+	jwt := getEnvVar("jwt_token2", "token ID")
 
 	ko := options.KeyOpts{
 		SkipConfirmation: true,
@@ -232,6 +170,76 @@ func TestCodeImageAndVerifyKeyless(t *testing.T) {
 		RekorURL:         options.DefaultRekorURL,
 	}
 	err := s.SignCmd(ro, ko, options.RegistryOptions{}, nil, []string{imageUri}, "", "", true, "", "", "", false, false, "", false)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	functionArn, err := createImageLambda(t)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	successTagValue := "Function signed and verified"
+	success, timeout := findTag(t, functionArn, lambdaClient, "Function clarity result", successTagValue)
+	if timeout {
+		t.Fatal("test failed on timout, the required tag not added in the time period")
+	}
+	if !success {
+		t.Fatal("test failure: no " + successTagValue + " tag in the signed function")
+	}
+	fmt.Println(successTagValue + " tag found in the signed function")
+	deleteLambda(imageFuncName)
+}
+
+func TestCodeSignAndVerify(t *testing.T) {
+	os.Setenv(integrity.ExperimentalEnv, "0")
+	viper.Set("privatekey", privateKey)
+	switchConfiguration(false, publicKey)
+
+	funcDefer, err := mockStdin(t, pass)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer funcDefer()
+
+	sbo := o.SignBlobOptions{
+		SignBlobOptions: options.SignBlobOptions{
+			Base64Output: true,
+			Registry:     options.RegistryOptions{},
+		},
+	}
+	err = sign.SignAndUploadCode(awsClient, "utils/testing_lambda", &sbo, ro)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	functionArn := initCodeLambda(t)
+
+	successTagValue := "Function signed and verified"
+	success, timeout := findTag(t, functionArn, lambdaClient, "Function clarity result", successTagValue)
+	if timeout {
+		t.Fatal("test failed on timout, the required tag not added in the time period")
+	}
+	if !success {
+		t.Fatal("test failure: no " + successTagValue + " tag in the signed function")
+	}
+	fmt.Println(successTagValue + " tag found in the signed function")
+	deleteLambda(codeFuncName)
+	deleteS3BucketContent(&bucket, []string{"function-clarity.zip"})
+}
+
+func TestImageSignAndVerify(t *testing.T) {
+	os.Setenv(integrity.ExperimentalEnv, "0")
+	switchConfiguration(false, publicKey)
+
+	funcDefer, err := mockStdin(t, pass)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer funcDefer()
+
+	ko := options.KeyOpts{KeyRef: privateKey, PassFunc: passFunc}
+	err = s.SignCmd(ro, ko, options.RegistryOptions{}, nil, []string{imageUri}, "", "", true, "", "", "", false, false, "", false)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -290,7 +298,7 @@ func findTag(t *testing.T, functionArn string, lambdaClient *lambda.Client, succ
 	return success, false
 }
 
-func switchConfigurationToKeyless() {
+func switchConfiguration(isKeyless bool, publicKey string) {
 	funcCfg, err := lambdaClient.GetFunctionConfiguration(context.TODO(), &lambda.GetFunctionConfigurationInput{FunctionName: aws.String(verifierFunctionName)})
 	if err != nil {
 		log.Fatal("failed to get function configuration")
@@ -305,8 +313,8 @@ func switchConfigurationToKeyless() {
 	if err != nil {
 		log.Fatal("failed to unmarshal config from yaml")
 	}
-	config.IsKeyless = true
-	config.PublicKey = ""
+	config.IsKeyless = isKeyless
+	config.PublicKey = publicKey
 
 	cfgYaml, err := yaml.Marshal(&config)
 	if err != nil {
@@ -322,6 +330,7 @@ func switchConfigurationToKeyless() {
 	if err != nil {
 		log.Fatalf("failed to update function configuration: %v", err)
 	}
+	time.Sleep(30 * time.Second)
 }
 
 func createConfig(region string) *aws.Config {
@@ -356,13 +365,13 @@ func deleteS3TrailBucketContent() {
 				continue
 			}
 			if string(bl.LocationConstraint) == region {
-				deleteS3BucketContent(bucket.Name)
+				deleteS3BucketContent(bucket.Name, []string{})
 			}
 		}
 	}
 }
 
-func deleteS3BucketContent(name *string) {
+func deleteS3BucketContent(name *string, except []string) {
 	listObjectsV2Response, err := s3Client.ListObjectsV2(context.TODO(),
 		&s3.ListObjectsV2Input{
 			Bucket: name,
@@ -374,13 +383,15 @@ func deleteS3BucketContent(name *string) {
 			log.Fatalf("Couldn't list objects... delete all objects in bucket: %s failed", *name)
 		}
 		for _, item := range listObjectsV2Response.Contents {
-			_, err = s3Client.DeleteObject(context.Background(), &s3.DeleteObjectInput{
-				Bucket: name,
-				Key:    item.Key,
-			})
+			if !contains(except, *item.Key) {
+				_, err = s3Client.DeleteObject(context.Background(), &s3.DeleteObjectInput{
+					Bucket: name,
+					Key:    item.Key,
+				})
 
-			if err != nil {
-				log.Fatalf("delete all objects in bucket: %s failed", *name)
+				if err != nil {
+					log.Fatalf("delete all objects in bucket: %s failed", *name)
+				}
 			}
 		}
 
@@ -396,11 +407,21 @@ func deleteS3BucketContent(name *string) {
 	}
 }
 
+func contains(s []string, str string) bool {
+	for _, v := range s {
+		if v == str {
+			return true
+		}
+	}
+
+	return false
+}
+
 func deleteS3Bucket(name string) {
 	if _, err := s3Client.HeadBucket(context.TODO(), &s3.HeadBucketInput{Bucket: aws.String(name)}); err != nil {
 		return
 	}
-	deleteS3BucketContent(&name)
+	deleteS3BucketContent(&name, []string{})
 	if _, err := s3Client.DeleteBucket(context.TODO(), &s3.DeleteBucketInput{Bucket: aws.String(name)}); err != nil {
 		log.Fatalf("delete bucket: %s failed", name)
 	}
@@ -439,7 +460,7 @@ func deleteStack() {
 		if timeout {
 			log.Fatal("timout on waiting for stack to delete")
 		}
-		time.Sleep(30 * time.Second)
+		time.Sleep(15 * time.Second)
 	}
 }
